backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
pydantic[email]==2.7.1
sendgrid==6.11.0
twilio==9.0.5
apscheduler==3.10.4
httpx==0.27.0
python-multipart==0.0.9
